1641: move test util into scripts
