+++ docker/dockerfiles/Dockerfile.server
@@ -45,7 +45,6 @@ COPY alembic.ini alembic.ini
 COPY codalab codalab
 COPY codalab_service.py codalab_service.py
 COPY docs docs
-COPY lib lib
 COPY mkdocs.yml mkdocs.yml
 COPY monitor.py monitor.py
 COPY scripts scripts
