+++ scripts/create-sample-worksheet.py
@@ -6,7 +6,7 @@ import sys
 import time
 
 sys.path.append('..')
-from codalab.lib.test_util import cleanup, run_command
+from scripts.test_util import cleanup, run_command
 
 """
 Script to create small and large sample worksheets in any instance to stress test the front end. The purpose of
